comment level access

+++ server/src/api/comments.ts
@@ -92,42 +92,42 @@ type CommentsSpreadsheetForamt = typeof commentsSpreadsheetFormats[number];
 const validCommentSpreadsheetFormat = (format: any): format is CommentsSpreadsheetForamt => commentsSpreadsheetFormats.includes(format);
 
 router
-	.all('*', (req, res, next) => {
-		const access = req.permissions?.comments || AccessLevel.none;
-		
-		if (req.method === "GET" && access >= AccessLevel.ro)
-			return next();
-		// Need read-only privileges to export comments
-		if (req.method === "POST" && req.path.search(/^\/export/i) >= 0 && access >= AccessLevel.ro)
-			return next();
-		// Need read-only privileges to update specific comments. We do a futher check for read-write on the specific comment later.
-		if (req.method === "PATCH" && req.path === '/' && access >= AccessLevel.ro)
-			return next();
-		// Need admin privileges to delete or add comments, or update start comment ID
-		if (['PATCH', 'DELETE', 'POST'].includes(req.method) && access >= AccessLevel.admin)
-			return next();
-		
-		next(new ForbiddenError('Insufficient karma'));
-	})
 	.patch('/startCommentId', (req, res, next) => {
+		const access = req.permissions?.comments || AccessLevel.none;
+		// Need admin privileges to change CIDs
+		if (access < AccessLevel.admin)
+			return next(new ForbiddenError("Need admin privileges at the ballot level to change CIDs"));
+
 		const ballot_id = req.ballot!.id;
 		if (!validUploadParams(req.body))
 			return next(new TypeError("Bad body; expected object with shape {startCommentId?: number}"));
 		const startCommentId = req.body.startCommentId || 1;
+
 		setStartCommentId(req.user, ballot_id, startCommentId)
 			.then(data => res.json(data))
 			.catch(next);
 	})
 	.post('/import', (req, res, next) => {
+		const access = req.permissions?.comments || AccessLevel.none;
+		// Need admin privileges for import
+		if (access < AccessLevel.admin)
+			return next(new ForbiddenError("Need admin privileges at the ballot level to import comments"));
+
 		const ballot = req.ballot!;
 		if (!validUploadParams(req.body))
 			return next(new TypeError("Bad body; expected object with shape {startCommentId?: number}"));
 		const startCommentId = req.body.startCommentId || 1;
+
 		importEpollComments(req.user, ballot, startCommentId)
 			.then(data => res.json(data))
 			.catch(next);
 	})
 	.post('/upload', upload.single('CommentsFile'), (req, res, next) => {
+		const access = req.permissions?.comments || AccessLevel.none;
+		// Need admin privileges for upload
+		if (access < AccessLevel.admin)
+			return next(new ForbiddenError("Need admin privileges at the ballot level to upload comments"));
+
 		let params: any;
 		try {
 			if (typeof req.body.params !== 'string')
@@ -149,6 +149,10 @@ router
 			.catch(next);
 	})
 	.post('/export', upload.single('file'), (req, res, next) => {
+		const access = req.permissions?.comments || AccessLevel.none;
+		if (access < AccessLevel.ro)
+			return next(new ForbiddenError("Need at least read-only privileges at the ballot level to export comments"));
+
 		let format: CommentsSpreadsheetForamt = "legacy";
 		if (typeof req.query.format === 'string') {
 			let formatIn = req.query.format.toLowerCase();
@@ -183,6 +187,10 @@ router
 	})
 	.route('/')
 		.get((req, res, next) => {
+			const access = req.permissions?.comments || AccessLevel.none;
+			if (access < AccessLevel.ro)
+				return next(new ForbiddenError("Need at least read-only privileges at the ballot level to get comments"));
+
 			const modifiedSince = typeof req.query.modifiedSince === 'string'? req.query.modifiedSince: undefined;
 			getComments(req.ballot!.id, modifiedSince)
 				.then(data => res.json(data))
@@ -190,15 +198,25 @@ router
 		})
 		.patch((req, res, next) => {
 			const access = req.permissions?.comments || AccessLevel.none;
+			// Need at least read-only privileges to update comments; check for comment level privileges later
+			if (access < AccessLevel.ro)
+				return next(new ForbiddenError("Need at least read-only privileges at the ballot level to update comments"));
+
 			const modifiedSince = typeof req.query.modifiedSince === 'string'? req.query.modifiedSince: undefined;
 			const updates = req.body;
 			if (!validUpdates(updates))
 				return next(new TypeError('Bad or missing updates; expected an array of objects with shape {id, changes}'));
+
 			updateComments(req.user, req.ballot!.id, access, updates, modifiedSince)
 				.then(data => res.json(data))
 				.catch(next);
 		})
 		.delete((req, res, next) => {
+			const access = req.permissions?.comments || AccessLevel.none;
+			// Need admin privileges to delete comments
+			if (access < AccessLevel.admin)
+				return next(new ForbiddenError("Need admin privileges at the ballot level to delete comments"));
+
 			deleteComments(req.user, req.ballot!.id)
 				.then(data => res.json(data))
 				.catch(next);
